+++ docs/changelog.md
@@ -41,6 +41,7 @@ The most exciting change is the ability to use blocks in multiple contexts, not 
 
 ### Fixed
 
+- New blocks are always saved in order, fixes [#75](https://github.com/nathanhollows/Rapua/issues/75).
 - Check in view now displays correct icons and long location names no longer look crowded.
 - Check out logic for tracking players at each location no longer triggers twice.
 - [Quiz block](/docs/user/blocks/quiz) correctly renders if the user revisits the location after having completed the quiz.